refactor(core-ui): extract shared UI into :apps:flipcash:core-ui
Move the shared Compose UI out of :apps:flipcash:core into a new
:apps:flipcash:core-ui module so it can depend on :shared:featureflags
(core cannot without a dependency cycle).

- Move all composables under core/ui (+ internal/ui/previews) to core-ui,
  keeping the com.flipcash.app.core.ui package so call-site imports are
  unchanged.
- Keep SteppedFlowRoute (a route type, not UI) in core, relocated to the
  core.navigation package.
- Wire core-ui into feature modules via the feature convention plugin,
  excluding core, core-ui and featureflags to avoid a cycle; add explicit
  deps for the two non-feature-plugin consumers (app, chat-ui).
