+++ apps/flipcash/core-ui/.gitignore
@@ -1,0 +1,2 @@
+build/
+.gradle/
